refactor(setup): split setup-wizard.ts into step modules
Extract 627-line monolith into focused modules:
- setup/utils.ts — guardCancel helper
- setup/steps-identity.ts — config check, user/agent name (81 lines)
- setup/steps-provider.ts — preset, provider, model, API key (179 lines)
- setup/steps-adapter.ts — adapter selection (29 lines)
- setup/steps-runtime.ts — daemon configuration (177 lines)

setup-wizard.ts is now a thin orchestrator (115 lines).
All 13 tests pass, build clean.
